refactor(preproc): modularize source provenance code
